requirements.txt
fastapi[all]
pillow 
tensorflow
keras
numpy
requests